Needs Analysis

+++ src/main/java/com/example/persona/repository/UserPersonaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserPersonaRepository extends JpaRepository<UserPersona, Integer> {
+public interface UserPersonaRepository extends JpaRepository<UserPersona, Long> {
     Optional<UserPersona> findByEmail(String email);
     Optional<UserPersona> findByUid(String uID);
     Optional<UserPersona> findByCid(Long cid);
